+++ client/src/components/Filters/PlatformFilter/index.js
@@ -3,7 +3,7 @@ import { useDispatch } from 'react-redux';
 import * as c from './index.module.css';
 
 
-import Toggle from '../Toggle';
+import Toggle from 'components/Toggle';
 
 const PlatformFilter = (props) => {
   const { platforms } = props;
@@ -20,7 +20,7 @@ const PlatformFilter = (props) => {
   }
 
   return (
-    <div className='Filter'>
+    <div className={c.Filter}>
       <h3>Platforms</h3>
       <Toggle name='Facebook' toggled={platforms.includes('facebook')} onToggled={onToggled('facebook')} />
       <Toggle name='Instagram' toggled={platforms.includes('instagram')} onToggled={onToggled('instagram')} />
