+++ client/src/components/Filters/PlatformFilter/index.module.css
@@ -1,0 +1,18 @@
+.Filters {
+  background-color: white;
+  border-right: 1px solid lightgray;
+  flex: 0 0 250px;
+  padding: 10px;
+  overflow-y: auto;
+  user-select: none;
+}
+
+.Filter {
+  background-color: white;
+  padding: 10px;
+  margin-bottom: 10px;
+}
+
+.Filter h3 {
+  margin: 0px 0px 10px 0px;
+}
