+++ client/src/components/OrganizationSettings/index.js
@@ -1,4 +1,4 @@
-import OrganizationModal from '../OrganizationModal';
+import OrganizationModal from './OrganizationModal';
 import * as c from './index.module.css';
